Separate OMR x86 linkage properties from linkage class
